feat: merge gmail_send_with_attachments into messaging_send
