+++ assistant/src/config/bundled-skills/gmail/TOOLS.json
@@ -349,55 +349,6 @@
       "execution_target": "host"
     },
     {
-      "name": "gmail_send_with_attachments",
-      "description": "Create a Gmail draft with file attachments for review. Include a confidence score (0-1).",
-      "category": "gmail",
-      "risk": "high",
-      "input_schema": {
-        "type": "object",
-        "properties": {
-          "to": {
-            "type": "string",
-            "description": "Recipient email address"
-          },
-          "subject": {
-            "type": "string",
-            "description": "Email subject line"
-          },
-          "body": {
-            "type": "string",
-            "description": "Email body (plain text)"
-          },
-          "attachment_paths": {
-            "type": "array",
-            "items": {
-              "type": "string"
-            },
-            "description": "Absolute file paths of attachments to include"
-          },
-          "in_reply_to": {
-            "type": "string",
-            "description": "Message-ID header for replies"
-          },
-          "thread_id": {
-            "type": "string",
-            "description": "Thread ID to continue an existing thread"
-          },
-          "confidence": {
-            "type": "number",
-            "description": "Confidence score (0-1) for this action"
-          },
-          "reason": {
-            "type": "string",
-            "description": "Brief non-technical explanation of why this tool is being called"
-          }
-        },
-        "required": ["to", "subject", "body", "attachment_paths", "confidence"]
-      },
-      "executor": "tools/gmail-send-with-attachments.ts",
-      "execution_target": "host"
-    },
-    {
       "name": "gmail_forward",
       "description": "Create a draft forwarding a Gmail message to another recipient, preserving attachments. Include a confidence score (0-1).",
       "category": "gmail",
