+++ assistant/src/config/bundled-skills/messaging/SKILL.md
@@ -155,7 +155,7 @@ When sending a Slack message, retain the `ts` (message timestamp) from the send 
 
 - **List Attachments**: `gmail_list_attachments` — list all attachments on a message with filename, MIME type, size, and attachment ID
 - **Download Attachment**: `gmail_download_attachment` — download an attachment to the working directory by message ID + attachment ID
-- **Send with Attachments**: `gmail_send_with_attachments` — send an email with file attachments (reads files from disk, builds multipart MIME)
+- **Send with Attachments**: `messaging_send` with `attachment_paths` — create a Gmail draft with file attachments (reads files from disk, builds multipart MIME)
 
 Workflow: use `gmail_list_attachments` to discover attachments, then `gmail_download_attachment` to save them locally.
 
@@ -217,7 +217,7 @@ Gmail uses a **draft-first workflow**. All compose and reply tools create Gmail 
 - `messaging_send` (Gmail) → creates a draft in Gmail Drafts
 - `messaging_reply` (Gmail) → creates a threaded draft with reply-all recipients
 - `gmail_draft` → creates a draft
-- `gmail_send_with_attachments` → creates a draft with attachments
+- `messaging_send` with `attachment_paths` → creates a draft with attachments
 - `gmail_forward` → creates a forward draft
 
 **To actually send**: Use `gmail_send_draft` with the draft ID after the user has reviewed it. Only call `gmail_send_draft` when the user explicitly says "send it" or equivalent.
