+++ assistant/src/config/bundled-skills/messaging/tools/messaging-send.ts
@@ -1,8 +1,16 @@
-import { createDraft } from "../../../../messaging/providers/gmail/client.js";
+import { readFile } from "node:fs/promises";
+import { basename } from "node:path";
+
+import {
+  createDraft,
+  createDraftRaw,
+} from "../../../../messaging/providers/gmail/client.js";
+import { buildMultipartMime } from "../../../../messaging/providers/gmail/mime-builder.js";
 import type {
   ToolContext,
   ToolExecutionResult,
 } from "../../../../tools/types.js";
+import { guessMimeType } from "./gmail-mime-helpers.js";
 import { err, ok, resolveProvider, withProviderToken } from "./shared.js";
 
 export async function run(
@@ -14,6 +22,8 @@ export async function run(
   const text = input.text as string;
   const subject = input.subject as string | undefined;
   const inReplyTo = input.in_reply_to as string | undefined;
+  const attachmentPaths = input.attachment_paths as string[] | undefined;
+  const threadId = input.thread_id as string | undefined;
 
   if (!conversationId) {
     return err("conversation_id is required.");
@@ -25,9 +35,41 @@ export async function run(
   try {
     const provider = resolveProvider(platform);
 
+    // Non-Gmail platforms: reject attachment_paths
+    if (provider.id !== "gmail" && attachmentPaths?.length) {
+      return err("Attachments are only supported on Gmail.");
+    }
+
     // Gmail: create a draft instead of sending directly
     if (provider.id === "gmail") {
       return withProviderToken(provider, async (token) => {
+        // With attachments: build multipart MIME and use createDraftRaw
+        if (attachmentPaths?.length) {
+          const attachments = await Promise.all(
+            attachmentPaths.map(async (filePath) => {
+              const data = await readFile(filePath);
+              const filename = basename(filePath);
+              const mimeType = guessMimeType(filePath);
+              return { filename, mimeType, data };
+            }),
+          );
+
+          const raw = buildMultipartMime({
+            to: conversationId,
+            subject: subject ?? "",
+            body: text,
+            inReplyTo,
+            attachments,
+          });
+          const draft = await createDraftRaw(token, raw, threadId);
+
+          const filenames = attachments.map((a) => a.filename).join(", ");
+          return ok(
+            `Gmail draft created with ${attachments.length} attachment(s): ${filenames} (Draft ID: ${draft.id}). Review in Gmail Drafts, then tell me to send it or send it yourself.`,
+          );
+        }
+
+        // Without attachments: use standard createDraft
         const draft = await createDraft(
           token,
           conversationId,
